fixed: wrong cast to int: Only major versions are taken into account (ie. Dataverse)

+++ application/modules/journal/controllers/SubmitController.php
@@ -376,7 +376,7 @@ class SubmitController extends DefaultController
     private function extractVersion(array &$params): void
     {
         $params['version'] = (isset($params['version']) && is_numeric($params['version']))
-            ? (int)$params['version']
+            ? (float)$params['version']
             : 1;
     }
